style #6: refactoring package
